♻️ refact(orm): Move session manager directly in repository
