+++ api/domain/task/models/enums.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class TaskStatus(str, Enum):
+    FAILURE = "FAILURE"
+    PENDING = "PENDING"
+    SUCCESS = "SUCCESS"
